adding locks

+++ src/main/java/com/corona/insights/scheduler/FileProcessScheduler.java
@@ -1,6 +1,7 @@
 package com.corona.insights.scheduler;
 
 import com.corona.insights.client.FileContainerClient;
+import com.corona.insights.dao.DaoUtils;
 import com.corona.insights.service.CoronaDataEnrichmentService;
 import com.corona.insights.service.CoronaETLProcessingService;
 import com.corona.insights.service.CoronaFileProcessingService;
@@ -20,35 +21,41 @@ public class FileProcessScheduler {
     private CoronaFileProcessingService coronaFileProcessingService;
     private CoronaETLProcessingService coronaETLProcessingService;
     private CoronaDataEnrichmentService coronaDataEnrichmentService;
+    private DaoUtils daoUtils;
 
-    public FileProcessScheduler(FileContainerClient fileContainerClient, CoronaFileProcessingService coronaFileProcessingService, CoronaETLProcessingService coronaETLProcessingService, CoronaDataEnrichmentService coronaDataEnrichmentService) {
+    public FileProcessScheduler(FileContainerClient fileContainerClient, CoronaFileProcessingService coronaFileProcessingService, CoronaETLProcessingService coronaETLProcessingService, CoronaDataEnrichmentService coronaDataEnrichmentService, DaoUtils daoUtils) {
         this.fileContainerClient = fileContainerClient;
         this.coronaFileProcessingService = coronaFileProcessingService;
         this.coronaETLProcessingService = coronaETLProcessingService;
         this.coronaDataEnrichmentService = coronaDataEnrichmentService;
+        this.daoUtils = daoUtils;
     }
 
     @Scheduled(cron = "0 0 */2 ? * *")
     public void pollForFiles() {
-        log.info("Executing FileProcessScheduler, fetching files to process");
-        List<File> filesToProcess = fileContainerClient.getFilesToProcess();
-        log.info("Fetched = {} files for processing", filesToProcess.size());
+        if(!daoUtils.isUpdateLocked()) {
+            daoUtils.acquireUpdateLock();
+            log.info("Executing FileProcessScheduler, fetching files to process");
+            List<File> filesToProcess = fileContainerClient.getFilesToProcess();
+            log.info("Fetched = {} files for processing", filesToProcess.size());
 
-        List<File> newFiles = coronaFileProcessingService.getNewFilesToProcess(filesToProcess);
-        log.info("Found = {} new files to process", newFiles.size());
+            List<File> newFiles = coronaFileProcessingService.getNewFilesToProcess(filesToProcess);
+            log.info("Found = {} new files to process", newFiles.size());
 
-        if(!newFiles.isEmpty()) {
-            for (File file : newFiles) {
-                try {
-                    File fileToProcess = fileContainerClient.markFileForProcessing(file);
-                    coronaFileProcessingService.processFile(fileToProcess);
-                    fileContainerClient.deleteFile(fileToProcess);
-                } catch (Exception e) {
-                    log.error("Exception while processing file = {}, exception = {}", file.getName(), e.getMessage());
+            if (!newFiles.isEmpty()) {
+                for (File file : newFiles) {
+                    try {
+                        File fileToProcess = fileContainerClient.markFileForProcessing(file);
+                        coronaFileProcessingService.processFile(fileToProcess);
+                        fileContainerClient.deleteFile(fileToProcess);
+                    } catch (Exception e) {
+                        log.error("Exception while processing file = {}, exception = {}", file.getName(), e.getMessage());
+                    }
                 }
+                //coronaDataEnrichmentService.enrich();
+                coronaETLProcessingService.process();
             }
-            //coronaDataEnrichmentService.enrich();
-            coronaETLProcessingService.process();
+            daoUtils.releaseUpdateLock();
         }
     }
